# api/openapi.yaml
openapi: 3.1.0
info:
  title: DokoSoko Control Plane API
  version: 4.0.0
  description: >-
    Control-plane contract for DokoSoko's core documentation, SDK references, credentials, recipes, reviewed tools, native tools, and MCP connector. Cookie-authenticated mutations
    also require CSRF validation. Public MCP is optional and read-only.
servers:
  - url: https://dokosoko.example.com
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/v1/products/{product_id}/analyses:
    parameters:
      - $ref: '#/components/parameters/ProductID'
    get:
      operationId: listIntegrationAnalyses
      responses:
        '200':
          description: Newest evidence-backed analyses first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationAnalysisList'
    post:
      operationId: analyseIntegration
      description: >-
        Builds a versioned plan from known DokoSoko evidence. Supply integration_id to scope evidence to one API's exact documentation, contracts, SDK references, authorization,
        identity, and tool bindings; omit it for deployment-wide analysis. AI extraction may enrich the plan but cannot invent references or hide unknowns.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                integration_id:
                  type: string
                  minLength: 1
      responses:
        '201':
          description: Analysis ready for human review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationAnalysis'
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /api/v1/products/{product_id}/analyses/{analysis_id}:
    parameters:
      - $ref: '#/components/parameters/ProductID'
      - name: analysis_id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getIntegrationAnalysis
      responses:
        '200':
          description: Analysis and its explicit unknowns.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationAnalysis'
    patch:
      operationId: answerIntegrationUnknowns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                answers:
                  type: object
                  additionalProperties:
                    type: string
                    maxLength: 8000
              required:
                - answers
      responses:
        '200':
          description: Analysis with supplied answers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationAnalysis'
        '422':
          $ref: '#/components/responses/Error'
  /api/v1/products/{product_id}/analyses/{analysis_id}/recipes:
    parameters:
      - $ref: '#/components/parameters/ProductID'
      - name: analysis_id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: generateRecipes
      description: >-
        Creates review-state Markdown recipes and durable job records. Supply the same integration_id used to create a scoped analysis; cross-integration generation is rejected.
        Blocking unknowns prevent generation.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                integration_id:
                  type: string
                  minLength: 1
      responses:
        '201':
          description: Generated recipes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
...
  /api/v1/products/{product_id}/recipes/{recipe_id}:
    parameters:
      - $ref: '#/components/parameters/ProductID'
      - name: recipe_id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getRecipe
      responses:
        '200':
          description: Recipe plus complete immutable revision history.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  recipe:
                    $ref: '#/components/schemas/Recipe'
                  revisions:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipeRevision'
                required:
                  - recipe
                  - revisions
    patch:
      operationId: updateRecipeMarkdown
      description: Creates a human-authored revision. References must resolve to known analyzed HTTPS sources.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeRequest'
      responses:
        '200':
          description: New review-state recipe revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '422':
          $ref: '#/components/responses/Error'
  /api/v1/products/{product_id}/recipes/{recipe_id}/rework:
    parameters:
      - $ref: '#/components/parameters/ProductID'
      - name: recipe_id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: reworkRecipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                instruction:
                  type: string
                  minLength: 1
                  maxLength: 8000
              required:
                - instruction
      responses:
        '200':
          description: New generated revision ready for review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '422':
          $ref: '#/components/responses/Error'
  /api/v1/products/{product_id}/recipes/{recipe_id}/approve:
    parameters:
      - $ref: '#/components/parameters/ProductID'
      - name: recipe_id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: approveRecipe
      description: Records explicit human approval of the current revision. Validation errors fail closed.
      responses:
        '200':
          description: Approved recipe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '422':
          $ref: '#/components/responses/Error'
  /api/v1/products/{product_id}/recipes/{recipe_id}/publish:
    parameters:
      - $ref: '#/components/parameters/ProductID'
      - name: recipe_id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: publishRecipe
      description: Publishes only the approved current revision. Public recipes may reference only public, published, non-quarantined sources.
      responses:
        '200':
          description: Recipe available as a stable MCP resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '422':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dokosoko_session
    bearerAuth:
      type: http
      scheme: bearer
      description: Administrator automation token in development only; production uses the root session cookie.
    setupToken:
      type: http
      scheme: bearer
      description: One-time deployment setup token.
    dokoBearer:
      type: http
      scheme: bearer
      description: Opaque, short-lived DokoSoko OAuth access token bound to one exact Private MCP resource.
  parameters:
    ProductID:
      name: product_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    IntegrationEvidence:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
        resource_id:
          type: string
        label:
          type: string
        location:
          type: string
        excerpt:
          type: string
          maxLength: 6000
          description: Bounded excerpt from published knowledge. Treat as untrusted data; sent only to the configured Analysis provider.
        references:
          type: array
          description: Exact canonical pages included in this evidence
          items:
            $ref: '#/components/schemas/RecipeReference'
        version:
          type: string
        visibility:
          $ref: '#/components/schemas/Visibility'
        fingerprint:
          type: string
          description: Exact content fingerprint used for drift detection.
      required:
        - kind
        - resource_id
        - label
        - visibility
        - fingerprint
    IntegrationUnknown:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        question:
          type: string
        why:
          type: string
        blocking:
          type: boolean
        answer:
          type: string
      required:
        - id
        - question
        - why
        - blocking
    IntegrationEndpointPlan:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        method:
          type: string
        path:
          type: string
        purpose:
          type: string
        identity:
          type: string
        evidence:
          type: array
          items:
            type: string
      required:
        - name
        - method
        - path
        - purpose
        - identity
        - evidence
    IntegrationIdentityPlan:
      type: object
      additionalProperties: false
      properties:
        mode:
          type: string
        issuer:
          type: string
          format: uri
        audience:
          type: string
        grants:
          type: array
          items:
            type: string
        explanation:
          type: string
      required:
        - mode
        - explanation
    RecipeSeed:
      type: object
      additionalProperties: false
      properties:
        slug:
          type: string
        title:
          type: string
        outcome:
          type: string
        audience:
          type: string
        endpoint_ids:
          type: array
          items:
            type: string
      required:
        - slug
        - title
        - outcome
        - audience
    IntegrationPlan:
      type: object
      additionalProperties: false
      properties:
        summary:
          type: string
        identity:
          $ref: '#/components/schemas/IntegrationIdentityPlan'
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationEndpointPlan'
        recipes:
          type: array
          items:
            $ref: '#/components/schemas/RecipeSeed'
      required:
        - summary
        - identity
        - endpoints
        - recipes
    IntegrationAnalysis:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        organisation_id:
          type: string
        product_id:
          type: string
        schema_version:
          type: integer
          const: 1
        state:
          type: string
          enum:
            - running
            - review
            - failed
        generated_by:
          type: string
          enum:
            - deterministic
            - ai_assisted
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationEvidence'
        plan:
          $ref: '#/components/schemas/IntegrationPlan'
        unknowns:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationUnknown'
        error_code:
          type: string
        revision:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
      required:
        - id
        - organisation_id
        - product_id
        - schema_version
        - state
        - generated_by
        - evidence
        - plan
        - unknowns
        - revision
        - created_at
    IntegrationAnalysisList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationAnalysis'
      required:
        - items
    RecipeReference:
      type: object
      additionalProperties: false
      properties:
        label:
          type: string
        url:
          type: string
          format: uri
          description: HTTPS source URL already present in the analysis evidence manifest.
        kind:
          type: string
        resource_id:
          type: string
        anchor:
          type: string
      required:
        - label
        - url
        - kind
    RecipeDependency:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
        resource_id:
          type: string
        version:
          type: string
          description: Opaque evidence fingerprint captured when this recipe revision was generated.
      required:
        - kind
        - resource_id
        - version
    RecipeValidationFinding:
      type: object
      additionalProperties: false
      properties:
        level:
          type: string
          enum:
            - info
            - warning
            - error
        code:
          type: string
        message:
          type: string
      required:
        - level
        - code
        - message
    RecipeRevision:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        recipe_id:
          type: string
        revision:
          type: integer
          minimum: 1
        markdown:
          type: string
        references:
          type: array
          items:
            $ref: '#/components/schemas/RecipeReference'
        validation:
          type: array
          items:
            $ref: '#/components/schemas/RecipeValidationFinding'
        review:
          type: string
        generated_by:
          type: string
          enum:
            - ai
            - human
            - deterministic
        model:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - recipe_id
        - revision
        - markdown
        - references
        - validation
        - generated_by
        - created_by
        - created_at
    Recipe:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        organisation_id:
          type: string
        product_id:
          type: string
        analysis_id:
          type: string
        slug:
          type: string
        title:
          type: string
        outcome:
          type: string
        audience:
          type: string
        state:
          type: string
          enum:
            - draft
            - review
            - approved
            - published
            - outdated
        generated:
          type: boolean
        needs_attention:
          type: boolean
        visibility:
          $ref: '#/components/schemas/Visibility'
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/RecipeDependency'
        current_revision_id:
          type: string
        current_revision:
          $ref: '#/components/schemas/RecipeRevision'
        stable_uri:
          type: string
          pattern: ^dokosoko://products/[^/]+/recipes/[^/]+$
        approved_by:
          type: string
        approved_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
        revision:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - organisation_id
        - product_id
        - slug
        - title
        - outcome
        - audience
        - state
        - generated
        - needs_attention
        - visibility
        - dependencies
        - current_revision_id
        - stable_uri
        - revision
        - created_at
        - updated_at
    RecipeList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Recipe'
      required:
        - items
    UpdateRecipeRequest:
      type: object
      additionalProperties: false
      properties:
        markdown:
          type: string
          minLength: 1
        references:
          type: array
          items:
            $ref: '#/components/schemas/RecipeReference'
        visibility:
          $ref: '#/components/schemas/Visibility'
      required:
        - markdown
        - references
        - visibility
...
    Visibility:
      type: string
      enum:
        - private
        - public
      default: private
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - 'null'
              additionalProperties: true
          required:
            - code
            - message
      required:
        - error
...